refactor(json-rpc): stop type drifting between client and server delegating the type definition to floresta-rpc
