+++ crates/floresta-node/src/json_rpc/blockchain.rs
@@ -12,18 +12,18 @@ use bitcoin::ScriptBuf;
 use bitcoin::Txid;
 use corepc_types::v29::GetTxOut;
 use corepc_types::ScriptPubkey;
+use floresta_rpc::rpc_types::GetBlockResVerbose;
+use floresta_rpc::rpc_types::GetBlockchainInfoRes;
+use floresta_rpc::rpc_types::GetTxOutProof;
+use floresta_rpc::rpc_types::RescanConfidence;
 use miniscript::descriptor::checksum;
 use serde_json::json;
 use serde_json::Value;
 use tracing::debug;
 
-use super::res::GetBlockResVerbose;
-use super::res::GetBlockchainInfoRes;
-use super::res::GetTxOutProof;
-use super::res::JsonRpcError;
 use super::server::RpcChain;
 use super::server::RpcImpl;
-use crate::json_rpc::res::RescanConfidence;
+use crate::json_rpc::res::JsonRpcError;
 
 impl<Blockchain: RpcChain> RpcImpl<Blockchain> {
     async fn get_block_inner(&self, hash: BlockHash) -> Result<Block, JsonRpcError> {
