+++ crates/floresta-rpc/src/rpc.rs
@@ -8,7 +8,14 @@ use serde_json::Number;
 use serde_json::Value;
 
 use crate::rpc_types;
-use crate::rpc_types::*;
+use crate::rpc_types::AddNodeCommand;
+use crate::rpc_types::Error;
+use crate::rpc_types::GetBlockRes;
+use crate::rpc_types::GetBlockchainInfoRes;
+use crate::rpc_types::GetMemInfoRes;
+use crate::rpc_types::GetRpcInfoRes;
+use crate::rpc_types::PeerInfo;
+use crate::rpc_types::RescanConfidence;
 
 type Result<T> = std::result::Result<T, rpc_types::Error>;
 
@@ -21,19 +28,23 @@ pub trait FlorestaRPC {
     /// for a given block height, encoded as a hexadecimal string.
     /// You need to have enabled block filters by setting the `blockfilters=1` option
     fn get_block_filter(&self, height: u32) -> Result<String>;
+
     /// Returns general information about the chain we are on
     ///
     /// This method returns a bunch of information about the chain we are on, including
     /// the current height, the best block hash, the difficulty, and whether we are
     /// currently in IBD (Initial Block Download) mode.
     fn get_blockchain_info(&self) -> Result<GetBlockchainInfoRes>;
+
     /// Returns the hash of the best (tip) block in the most-work fully-validated chain.
     fn get_best_block_hash(&self) -> Result<BlockHash>;
+
     /// Returns the hash of the block at the given height
     ///
     /// This method returns the hash of the block at the given height. If the height is
     /// invalid, an error is returned.
     fn get_block_hash(&self, height: u32) -> Result<BlockHash>;
+
     /// Returns the block header for the given block hash
     ///
     /// This method returns the block header for the given block hash, as defined
@@ -41,17 +52,20 @@ pub trait FlorestaRPC {
     /// the previous block hash, the merkle root, the timestamp, the difficulty target,
     /// and the nonce.
     fn get_block_header(&self, hash: BlockHash) -> Result<BlockHeader>;
+
     /// Gets a transaction from the blockchain
     ///
     /// This method returns a transaction that's cached in our wallet. If the verbosity flag is
     /// set to false, the transaction is returned as a hexadecimal string. If the verbosity
     /// flag is set to true, the transaction is returned as a json object.
     fn get_transaction(&self, tx_id: Txid, verbosity: Option<bool>) -> Result<Value>;
+
     /// Returns the proof that one or more transactions were included in a block
     ///
     /// This method returns the Merkle proof, showing that a transaction was included in a block.
     /// The pooof is returned as a vector hexadecimal string.
     fn get_txout_proof(&self, txids: Vec<Txid>, blockhash: Option<BlockHash>) -> Option<String>;
+
     /// Loads up a descriptor into the wallet
     ///
     /// This method loads up a descriptor into the wallet. If the rescan option is not None,
@@ -72,40 +86,47 @@ pub trait FlorestaRPC {
 
     /// Returns the current height of the blockchain
     fn get_block_count(&self) -> Result<u32>;
+
     /// Sends a hex-encoded transaction to the network
     ///
     /// This method sends a transaction to the network. The transaction should be encoded as a
     /// hexadecimal string. If the transaction is valid, it will be broadcast to the network, and
     /// return the transaction id. If the transaction is invalid, an error will be returned.
     fn send_raw_transaction(&self, tx: String) -> Result<Txid>;
+
     /// Gets the current accumulator for the chain we're on
     ///
     /// This method returns the current accumulator for the chain we're on. The accumulator is
     /// a set of roots, that let's us prove that a UTXO exists in the chain. This method returns
     /// a vector of hexadecimal strings, each of which is a root in the accumulator.
     fn get_roots(&self) -> Result<Vec<String>>;
+
     /// Gets information about the peers we're connected with
     ///
     /// This method returns information about the peers we're connected with. This includes
     /// the peer's IP address, the peer's version, the peer's user agent, the transport protocol
     /// and the peer's current height.
     fn get_peer_info(&self) -> Result<Vec<PeerInfo>>;
+
     /// Returns a block, given a block hash
     ///
     /// This method returns a block, given a block hash. If the verbosity flag is 0, the block
     /// is returned as a hexadecimal string. If the verbosity flag is 1, the block is returned
     /// as a json object.
     fn get_block(&self, hash: BlockHash, verbosity: Option<u32>) -> Result<GetBlockRes>;
+
     /// Return a cached transaction output
     ///
     /// This method returns a cached transaction output. If the output is not in the cache,
     /// or is spent, an empty object is returned. If you want to find a utxo that's not in
     /// the cache, you can use the findtxout method.
     fn get_tx_out(&self, tx_id: Txid, outpoint: u32) -> Result<GetTxOut>;
+
     /// Stops the florestad process
     ///
     /// This can be used to gracefully stop the florestad process.
     fn stop(&self) -> Result<String>;
+
     /// Tells florestad to connect with a peer
     ///
     /// You can use this to connect with a given node, providing it's IP address and port.
@@ -113,6 +134,7 @@ pub trait FlorestaRPC {
     /// P2P protocol.
     #[doc = include_str!("../../../doc/rpc/addnode.md")]
     fn add_node(&self, node: String, command: AddNodeCommand, v2transport: bool) -> Result<Value>;
+
     /// Finds an specific utxo in the chain
     ///
     /// You can use this to look for a utxo. If it exists, it will return the amount and
@@ -125,16 +147,21 @@ pub trait FlorestaRPC {
         script: String,
         height_hint: u32,
     ) -> Result<Value>;
+
     /// Returns statistics about Floresta's memory usage.
     ///
     /// Returns zeroed values for all runtimes that are not *-gnu or MacOS.
     fn get_memory_info(&self, mode: String) -> Result<GetMemInfoRes>;
+
     /// Returns stats about our RPC server
     fn get_rpc_info(&self) -> Result<GetRpcInfoRes>;
+
     /// Returns for how long florestad has been running, in seconds
     fn uptime(&self) -> Result<u32>;
+
     /// Returns a list of all descriptors currently loaded in the wallet
     fn list_descriptors(&self) -> Result<Vec<String>>;
+
     /// Sends a ping to all peers, checking if they are still alive
     fn ping(&self) -> Result<()>;
 }
